refactor(tilrettelegging): ekstraher ArbeidsforholdHeader og bruk useWatch
- Ekstraher ArbeidsforholdHeader til egen komponent fra ArbeidsforholdExpansionCard
- Erstatt formMethods.watch() med useWatch for bedre ytelse (unngår unødvendige renderinger)
- Fiks import-sti for faisuUtils (fjern .tsx-endelse)
